Pin the PCA projection layer to float32
The pca_projection layer computes raw_counts @ loadings + bias, where
the two terms are large and nearly cancel to the much smaller PCA
score. Under the default mixed_float16 policy that cancellation loses
most of the result, and k-fold error collapses to the no-PCA baseline.
Pinning the layer to float32 keeps the projection exact while the rest
of the network still trains in mixed precision.

+++ locator/models.py
@@ -165,10 +165,15 @@ def create_network(
     # Optional PCA-initialized linear projection as the first layer. Placed
     # before BatchNormalization so it sees raw genotype counts, which lets a
     # caller set its weights to PCA loadings and reproduce PCA scores exactly.
+    # Pinned to float32: it computes raw_counts @ loadings + bias, where the
+    # two terms are large and nearly cancel, so float16 loses the result.
     if pca_components is not None:
-        x = layers.Dense(pca_components, activation="linear", name=PCA_LAYER_NAME)(
-            inputs
-        )
+        x = layers.Dense(
+            pca_components,
+            activation="linear",
+            name=PCA_LAYER_NAME,
+            dtype="float32",
+        )(inputs)
         x = layers.BatchNormalization()(x)
     else:
         x = layers.BatchNormalization()(inputs)
